fix: avoid code duplication with certificates.
fixes gravitee-io/issues#5925
