+++ gravitee-am-certificate/gravitee-am-certificate-api/pom.xml
@@ -45,5 +45,12 @@
             <groupId>io.reactivex.rxjava2</groupId>
             <artifactId>rxjava</artifactId>
         </dependency>
+
+        <!-- Nimbus jose+jwt -->
+        <dependency>
+            <groupId>com.nimbusds</groupId>
+            <artifactId>nimbus-jose-jwt</artifactId>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 </project>
